% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{Normalize a matrix or vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{a matrix or vector}
}
\value{
returns a normalized matrix or vector
}
\description{
Normalize values between 0 and 1, internal function only.
}
