% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_normalize.R
\name{raster_normalize}
\alias{raster_normalize}
\title{Normalize a raster layer}
\usage{
raster_normalize(x)
}
\arguments{
\item{x}{a raster layer}
}
\value{
returns a normalized matrix or vector
}
\description{
Normalize rasterlayer values between 0 and 1
}
\examples{

\donttest{
# fake raster data
x <- raster::raster(matrix(1:10, ncol = 2))
x <- raster_normalize(x)
}
}
\keyword{foto,}
\keyword{normalization}
\keyword{radial}
\keyword{spectrum,}
