% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspectrum.R
\name{rspectrum}
\alias{rspectrum}
\title{Calculates a radial spectrum}
\usage{
rspectrum(x, w, n = TRUE, env, ...)
}
\arguments{
\item{x}{a square matrix}

\item{w}{a moving window size}

\item{n}{normalize, bolean \code{TRUE} or \code{FALSE}}

\item{env}{local environment to evaluate}

\item{...}{additional parameters to forward}
}
\value{
returns a radial spectrum value for the matrix
}
\description{
This is an internal function and not to be used stand-alone.
}
\seealso{
\code{\link[foto]{foto}}
}
\keyword{foto,}
\keyword{radial}
\keyword{spectrum}
