% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspectrum.R
\name{rspectrum}
\alias{rspectrum}
\title{Calculates a radial spectrum}
\usage{
rspectrum(x, w, n = TRUE, h = TRUE, env, ...)
}
\arguments{
\item{x}{a square matrix}

\item{w}{a moving window size}

\item{n}{normalize, bolean \code{TRUE} or \code{FALSE}}

\item{h}{high pass filter on the two first spectra values
set to 0, limits the influence of low frequency components
bolean \code{TRUE} or \code{FALSE}}

\item{env}{local environment to evaluate}

\item{...}{additional parameters to forward}
}
\value{
Returns a radial spectrum values for the image used
in order to classify texture using a PCA (or other) analysis.
}
\description{
This is an internal function and not to be used stand-alone.
}
